// src/components/PdfEditor/types.ts
export type ReportBroTemplate = {
  docElements: {
    elementType: string;
    id: number;
    containerId: string;
    x: number;
    y: number;
    width: number;
    height: number;
    content: string;
    eval: boolean;
    styleId: string;
    bold: boolean;
    italic: boolean;
    underline: boolean;
    horizontalAlignment: string;
    verticalAlignment: string;
    textColor: string;
    backgroundColor: string;
    font: string;
    fontSize: number;
    lineSpacing: number;
    borderColor: string;
    borderWidth: number;
    borderAll: boolean;
    borderLeft: boolean;
    borderTop: boolean;
    borderRight: boolean;
    borderBottom: boolean;
    paddingLeft: number;
    paddingTop: number;
    paddingRight: number;
    paddingBottom: number;
    printIf: string;
    removeEmptyElement: boolean;
    alwaysPrintOnSamePage: boolean;
    pattern: string;
    cs_condition: string;
    cs_styleId: string;
    cs_bold: boolean;
    cs_italic: boolean;
    cs_underline: boolean;
    cs_horizontalAlignment: string;
    cs_verticalAlignment: string;
    cs_textColor: string;
    cs_backgroundColor: string;
    cs_font: string;
    cs_fontSize: number;
    cs_lineSpacing: number;
    cs_borderColor: string;
    cs_borderWidth: string;
    cs_borderAll: boolean;
    cs_borderLeft: boolean;
    cs_borderTop: boolean;
    cs_borderRight: boolean;
    cs_borderBottom: boolean;
    cs_paddingLeft: number;
    cs_paddingTop: number;
    cs_paddingRight: number;
    cs_paddingBottom: number;
    spreadsheet_hide: boolean;
    spreadsheet_column: string;
    spreadsheet_colspan: string;
    spreadsheet_addEmptyRow: boolean;
  }[];
  parameters: {
    id: number;
    name: string;
    type: 'string' | 'number' | 'boolean' | 'date' | 'image' | 'array' | 'simple_array' | 'average';
    arrayItemType: 'string';
    eval: boolean;
    nullable: boolean;
    pattern: string;
    expression: string;
    showOnlyNameType: boolean;
    testData: string;
  }[];
  styles: unknown[];
  version: number;
  documentProperties: {
    pageFormat: string;
    pageWidth: string;
    pageHeight: string;
    unit: string;
    orientation: 'landscape' | 'portrait';
    contentHeight: string;
    marginLeft: string;
    marginTop: string;
    marginRight: string;
    marginBottom: string;
    header: boolean;
    headerSize: string;
    headerDisplay: string;
    footer: boolean;
    footerSize: string;
    footerDisplay: string;
    patternLocale: string;
    patternCurrencySymbol: string;
  };
};